feat(gc): add faster GC option
Running `readdir("/nix/store")` is usually slower than asking the
SQLite3 database for those same store paths. That's what the GC does
by default because it also wants to cache incomplete builds that haven't
been registered in the SQLite3 database yet.

This commit adds `--ignore-unregistered-paths` to the GC options in
order to take the fast path. This allows the users to do a faster GC run
with the given trade-off. The idea would be that the users still does
the slow GC during the weekend / slow times.